logs/pyproject.toml
[tool.poetry]
name = "log-processor"
version = "0.1.0"
description = "Log processing service with API"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
bytewax = "^0.21.0"
sqlalchemy = "^2.0.0"
numpy = "^1.24.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
